requirements.txt
falcon==2.0.0
gunicorn==20.0.4
pyDataverse==0.2.1
python-dotenv==0.13.0
pyvips==2.1.12
requests==2.23.0